requirements.txt
requests==2.25.1
beautifulsoup4==4.10.0
bs4==0.0.1